+++ lib/active_delivery/lines/mailer.rb
@@ -5,7 +5,7 @@ module ActiveDelivery
     class Mailer < Base
       alias_method :mailer_class, :handler_class
 
-      DEFAULT_RESOLVER = ->(name) { name.gsub(/Delivery$/, "Mailer").safe_constantize }
+      DEFAULT_RESOLVER = ->(name) { name&.gsub(/Delivery$/, "Mailer")&.safe_constantize }
 
       def notify?(method_name)
         mailer_class.action_methods.include?(method_name.to_s)